[EH] make trait of default and hint
