+++ src/Contracts/CrudFieldAbstract.php
@@ -9,32 +9,4 @@ abstract class CrudFieldAbstract extends CrudAbstract implements CrudFieldInterf
     | Methods available for all Field Types
     |--------------------------------------------------------------------------
     */
-
-    public function prefix(string $prefix)
-    {
-        $this->options['prefix'] = $prefix;
-
-        return $this;
-    }
-
-    public function suffix(string $suffix)
-    {
-        $this->options['suffix'] = $suffix;
-
-        return $this;
-    }
-
-    public function default(string $default)
-    {
-        $this->options['default'] = $default;
-
-        return $this;
-    }
-
-    public function hint(string $hint)
-    {
-        $this->options['hint'] = $hint;
-
-        return $this;
-    }
 }
